fix: type of total_play_time is bigint

+++ structs/player.go
@@ -7,10 +7,10 @@ type Player struct {
 	uuid                 string    `gorm:"type:char(36);unique;not null"`
 	initial_login_date   time.Time `gorm:"not null"`
 	last_login_date      time.Time
-	total_play_time      int   `gorm:"default:0"`
-	experience           float `gorm:"default:0.0"`
-	currency             int   `gorm:"default:0"`
-	total_build_blocks   int   `gorm:"default:0"`
-	total_destroy_blocks int   `gorm:"default:0"`
-	total_mob_kills      int   `gorm:"default:0"`
+	total_play_time      bigint `gorm:"default:0"`
+	experience           float  `gorm:"default:0.0"`
+	currency             int    `gorm:"default:0"`
+	total_build_blocks   int    `gorm:"default:0"`
+	total_destroy_blocks int    `gorm:"default:0"`
+	total_mob_kills      int    `gorm:"default:0"`
 }
